requirements.txt
# Framework FastAPI e dependências
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
aiohttp==3.9.3

# Cliente OpenAI
openai==1.12.0

# Bibliotecas para teste
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-env==1.1.3

# Formatação e tipagem
black==24.1.1
mypy==1.8.0

# Utilitários
python-multipart==0.0.9